# src/configs/swagger-output.json
{
  "swagger": "2.0",
  "openapi": "3.0.0",
  "info": {
    "title": "MyBundee Auxiliary APIs",
    "version": "2.0.0",
    "description": "Auxiliary APIs for MyBundee. Includes chat, turo vehicle data, firebase admin operations and others."
  },
  "servers": [
    {
      "url": "https://bundee-auxiliary-services-dev.azurewebsites.net/",
      "description": "Deployed DEV on AZURE Environment"
    },
    {
      "url": "https://bundee-auxiliary-services-qa.azurewebsites.net/",
      "description": "Deployed QA on AZURE Environment"
    },
    {
      "url": "http://localhost:8000",
      "description": "Dev Localhost"
    }
  ],
  "paths": {
...
    "/api/v1/insurance/datarequests/new": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/insurance/datarequests/generate_invitation_link": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/insurance/datarequests/get_items": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "datarequestId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}